renamed old to browser file, moved all browser related stuff into browers file, server files are just in src, removed unused references
