+++ clientMod.ts
@@ -2,6 +2,6 @@ import {
   ObsidianWrapper,
   useObsidian,
 } from './ObsidianWrapper/ObsidianWrapper.jsx';
-import BrowserCache from './src/CacheClassBrowser.js';
+import BrowserCache from './src/Browser/CacheClassBrowser.js/index.js';
 
 export { BrowserCache, ObsidianWrapper, useObsidian };
