+++ src/Browser/normalize.js
