+++ src/astNormalize.js
@@ -3,7 +3,6 @@
 import { gql } from "https://deno.land/x/oak_graphql/mod.ts";
 import { print, visit } from "https://deno.land/x/graphql_deno/mod.ts";
 import { redisdb } from "./quickCache.js";
-import testsObj from "../queries.js";
 
 //graphql response is going to be in JSON;
 // this is for breaking up AST feilds/parts into the hash
@@ -20,13 +19,13 @@ const cacheWriteList = async (hash, array, overwrite = true) => {
   return;
 };
 
-const cacheReadList = async (hash) => {
-  let redisList = await redisdb.lrange(hash, 0, -1);
+// const cacheReadList = async (hash) => {
+//   let redisList = await redisdb.lrange(hash, 0, -1);
 
-  let cachedArray = redisList.map((element) => JSON.parse(element));
+//   let cachedArray = redisList.map((element) => JSON.parse(element));
 
-  return cachedArray;
-};
+//   return cachedArray;
+// };
 
 const cacheWriteObject = async (hash, obj) => {
   let entries = Object.entries(obj).flat();
@@ -35,28 +34,28 @@ const cacheWriteObject = async (hash, obj) => {
   await redisdb.hset(hash, ...entries);
 };
 
-const cacheReadObject = async (hash, field) => {
-  if (field) {
-    let returnValue = await redisdb.hget(hash, JSON.stringify(field));
+// const cacheReadObject = async (hash, field) => {
+//   if (field) {
+//     let returnValue = await redisdb.hget(hash, JSON.stringify(field));
 
-    if (returnValue === undefined) return undefined;
-    return JSON.parse(returnValue);
-  } else {
-    let objArray = await redisdb.hgetall(hash);
-    if (objArray.length == 0) return undefined;
-    let parsedArray = objArray.map((entry) => JSON.parse(entry));
+//     if (returnValue === undefined) return undefined;
+//     return JSON.parse(returnValue);
+//   } else {
+//     let objArray = await redisdb.hgetall(hash);
+//     if (objArray.length == 0) return undefined;
+//     let parsedArray = objArray.map((entry) => JSON.parse(entry));
 
-    if (parsedArray.length % 2 !== 0) {
-      return undefined;
-    }
-    let returnObj = {};
-    for (let i = 0; i < parsedArray.length; i += 2) {
-      returnObj[parsedArray[i]] = parsedArray[i + 1];
-    }
+//     if (parsedArray.length % 2 !== 0) {
+//       return undefined;
+//     }
+//     let returnObj = {};
+//     for (let i = 0; i < parsedArray.length; i += 2) {
+//       returnObj[parsedArray[i]] = parsedArray[i + 1];
+//     }
 
-    return returnObj;
-  }
-};
+//     return returnObj;
+//   }
+// };
 
 export async function normalizeResult(
   gqlResponse,
