+++ test_files/rhum_test_files/insertTypenames_test.ts
@@ -2,7 +2,7 @@ import {
   insertTypenames,
   addTypenamesToFieldsStr,
   findClosingBrace,
-} from '../../src/insertTypenames.js';
+} from '../../src/Browser/insertTypenames.js/index.js';
 import { Rhum } from 'https://deno.land/x/rhum@v1.1.4/mod.ts';
 import { test } from '../test_variables/insertTypenames_variables.ts';
