+++ test_files/rhum_test_files/lfuBrowserCache_test.ts
@@ -8,10 +8,10 @@
  * Should test capacity 
  */
 
-import LFUCache from '../../src/lfuBrowserCache.js';
+import LFUCache from '../../src/Browser/lfuBrowserCache.js/index.js';
 import { Rhum } from 'https://deno.land/x/rhum@v1.1.4/mod.ts';
 import { test } from '../test_variables/lfuBrowserCache_variables.ts';
-import normalizeResult from '../../src/Old/normalize.js';
+import normalizeResult from '../../src/Browser/normalize.js';
 
 Rhum.testPlan('LFU Browser Cache Testing', () => {
   Rhum.testSuite('write/read nested data object', () => {
