+++ test_files/rhum_test_files/writeCache_test.ts
@@ -8,7 +8,7 @@
  *
  */
 
-import Cache from '../../src/CacheClassBrowser.js';
+import Cache from '../../src/Browser/CacheClassBrowser.js/index.js';
 import { Cache as CacheServer } from '../../src/CacheClassServer.js';
 import { Rhum } from 'https://deno.land/x/rhum@v1.1.4/mod.ts';
 import { test } from '../test_variables/writeCache_variables.ts';
